bug 23783 Oops, I broke Windows.  srand needs an explicit (unsigned int) cast for its seed
git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@21191 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ components/devices/device/src/sbDeviceEnsureSpaceForWrite.cpp
@@ -196,7 +196,7 @@ sbDeviceEnsureSpaceForWrite::RemoveExtraItems() {
   // Seed our random number generator
   time_t t;
   time(&t);
-  srand(t);
+  srand((unsigned int) t);
 
   // Now we consider our ADDED changes in random order
   PRUint32 addChangesLength;
